Dynamic color selection of sprite sheet
